Add files via upload
The internet of things (IoT) has the potential to greatly impact our social, economic, and commercial lives, but the interconnectedness of devices and the limited computational power of IoT devices puts them at risk for security threats. Machine learning (ML) models have been proposed as a solution to provide security to IoT systems, as they can be trained on the large amounts of data generated by IoT devices. This work provides an overview of research in the field of IoT security and ML implementation, including the security issues in IoT networks, the advantages and disadvantages of different ML algorithms and models, and their current and potential applications in IoT. Limitations of using ML models to secure IoT networks are also discussed as potential future research directions.